wip: add all sw_dp_pio.c functions to probe.rs
Also take in some DAP constants that are required.

Think about how we are going to build the DAP.rs file.
It seems the lower level probe needs some information from a dap_data
structure that we have to manage somewhere else. It doesn't seem that
the probe needs to write to this data structure, but only read from it.