feat: Initial compiling and running bootloader
Need to be tested thoroughly but seems to boot. The application runs,
but if the bootloader is flashed with the release builds, watchdog will
be enabled and I'm not sure yet how I can easily debug the app after the
watchdog is enabled in the bootloader.

For now debug builds just build without the watchdog option.